+++ web/js/views/allboards.js
@@ -5,7 +5,7 @@ import { api } from '../api.js';
 import { icons, STATUS_META, STATUSES } from '../icons.js';
 import { esc } from '../util.js';
 import { cardTile, statusPill, openCreateDialog, openBoardDialog, openNewMenu, crumb } from '../components.js';
-import { tabs, needYouCount, openCount } from './board.js';
+import { tabs, wireTabs, needYouCount, openCount } from './board.js';
 
 export const boardDot = (i) => (i === 0 ? 'var(--brand)' : 'var(--mut-2)');
 
@@ -66,6 +66,7 @@ export async function renderAllBoards(root, { boards }) {
         .join('')}
     </div>
   `;
+  wireTabs(root);
   const goToCard = (card) => (location.hash = `#/card/${card.id}`);
   const goToBoard = (board) => (location.hash = `#/board/${board.id}`);
   root.querySelectorAll('[data-new-menu]').forEach((b) => {